+++ Assets/FrogFriction.physicsMaterial2D
@@ -6,9 +6,9 @@ PhysicsMaterial2D:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: FloorFriction
+  m_Name: FrogFriction
   serializedVersion: 2
-  friction: 5
+  friction: 10
   bounciness: 0
   m_FrictionCombine: 1
   m_BounceCombine: 4
