+++ Assets/Scripts/Camera.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 568d1352604b4f8a8a820793bc8b734b
+timeCreated: 1747973105
