+++ Assets/Sprites.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: a3276031337f2ff4d99e1e30af41a4b7
-NativeFormatImporter:
+guid: 35d28e04000d34944b8d2085fba05953
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 6200000
   userData: 
   assetBundleName: 
   assetBundleVariant:
