--- Assets/kuvitoenmaincra.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0b86dd02c0325be55b08126d1f9fb6c3
